refactor: move carbon estimation treemap component to components
