+++ src/app/carbon-estimation/carbon-estimation.component.ts
@@ -15,7 +15,7 @@ import {
 import { ExpansionPanelComponent } from '../components/expansion-panel/expansion-panel.component';
 import { TabsComponent } from '../components/tab/tabs/tabs.component';
 import { TabItemComponent } from '../components/tab/tab-item/tab-item.component';
-import { CarbonEstimationTreemapComponent } from '../carbon-estimation-treemap/carbon-estimation-treemap.component';
+import { CarbonEstimationTreemapComponent } from '../components/carbon-estimation-treemap/carbon-estimation-treemap.component';
 import { CarbonEstimation, EstimatorValues, JsonExport } from '../types/carbon-estimator';
 import { sumValues } from '../utils/number-object';
 import { estimatorHeights } from './carbon-estimation.constants';
