+++ src/app/components/export-modal/export-modal.component.spec.ts
@@ -3,7 +3,7 @@ import { ExportModal } from './export-modal.component';
 import { CarbonEstimation } from '../../types/carbon-estimator';
 import { FormsModule } from '@angular/forms';
 import { InputGroupDisplay } from '../input-group-display/input-group-display.component';
-import { CarbonEstimationTreemapComponent } from '../../carbon-estimation-treemap/carbon-estimation-treemap.component';
+import { CarbonEstimationTreemapComponent } from '../carbon-estimation-treemap/carbon-estimation-treemap.component';
 import { CarbonEstimationTableComponent } from '../../carbon-estimation-table/carbon-estimation-table.component';
 import { inputBinding } from '@angular/core';
 import { DisclaimerTextComponent } from '../../components/disclaimer-text/disclaimer-text.component';
